+++ src/PhpSpreadsheet/Reader/Xlsx.php
@@ -231,7 +231,6 @@ class Xlsx extends BaseReader
                 if ($xmlWorkbook->sheets) {
                     $dir = dirname($relTarget);
 
-                    /** @var SimpleXMLElement $eleSheet */
                     foreach ($xmlWorkbook->sheets->sheet as $eleSheet) {
                         $tmpInfo = [
                             'worksheetName' => (string) self::getAttributes($eleSheet)['name'],
@@ -743,7 +742,6 @@ class Xlsx extends BaseReader
                     $charts = $chartDetails = [];
 
                     if ($xmlWorkbookNS->sheets) {
-                        /** @var SimpleXMLElement $eleSheet */
                         foreach ($xmlWorkbookNS->sheets->sheet as $eleSheet) {
                             $eleSheetAttr = self::getAttributes($eleSheet);
                             ++$oldSheetId;
@@ -1428,14 +1426,17 @@ class Xlsx extends BaseReader
                                             foreach ($xmlDrawingChildren->oneCellAnchor as $oneCellAnchor) {
                                                 $oneCellAnchor = self::testSimpleXml($oneCellAnchor);
                                                 if ($oneCellAnchor->pic->blipFill) {
-                                                    /** @var SimpleXMLElement $blip */
+                                                    $objDrawing = new \PhpOffice\PhpSpreadsheet\Worksheet\Drawing();
                                                     $blip = $oneCellAnchor->pic->blipFill->children(Namespaces::DRAWINGML)->blip;
-                                                    /** @var SimpleXMLElement $xfrm */
+                                                    if (isset($blip, $blip->alphaModFix)) {
+                                                        $temp = (string) $blip->alphaModFix->attributes()->amt;
+                                                        if (is_numeric($temp)) {
+                                                            $objDrawing->setOpacity((int) $temp);
+                                                        }
+                                                    }
                                                     $xfrm = $oneCellAnchor->pic->spPr->children(Namespaces::DRAWINGML)->xfrm;
-                                                    /** @var SimpleXMLElement $outerShdw */
                                                     $outerShdw = $oneCellAnchor->pic->spPr->children(Namespaces::DRAWINGML)->effectLst->outerShdw;
 
-                                                    $objDrawing = new \PhpOffice\PhpSpreadsheet\Worksheet\Drawing();
                                                     $objDrawing->setName(self::getArrayItemString(self::getAttributes($oneCellAnchor->pic->nvPicPr->cNvPr), 'name'));
                                                     $objDrawing->setDescription(self::getArrayItemString(self::getAttributes($oneCellAnchor->pic->nvPicPr->cNvPr), 'descr'));
                                                     $embedImageKey = self::getArrayItemString(
@@ -1498,7 +1499,6 @@ class Xlsx extends BaseReader
                                                     $height = Drawing::EMUToPixels(self::getArrayItemIntOrSxml(self::getAttributes($oneCellAnchor->ext), 'cy'));
 
                                                     $graphic = $oneCellAnchor->graphicFrame->children(Namespaces::DRAWINGML)->graphic;
-                                                    /** @var SimpleXMLElement $chartRef */
                                                     $chartRef = $graphic->graphicData->children(Namespaces::CHART)->chart;
                                                     $thisChart = (string) self::getAttributes($chartRef, $xmlNamespaceBase);
 
@@ -1520,6 +1520,12 @@ class Xlsx extends BaseReader
                                                 if ($twoCellAnchor->pic->blipFill) {
                                                     $objDrawing = new \PhpOffice\PhpSpreadsheet\Worksheet\Drawing();
                                                     $blip = $twoCellAnchor->pic->blipFill->children(Namespaces::DRAWINGML)->blip;
+                                                    if (isset($blip, $blip->alphaModFix)) {
+                                                        $temp = (string) $blip->alphaModFix->attributes()->amt;
+                                                        if (is_numeric($temp)) {
+                                                            $objDrawing->setOpacity((int) $temp);
+                                                        }
+                                                    }
                                                     if (isset($twoCellAnchor->pic->blipFill->children(Namespaces::DRAWINGML)->srcRect)) {
                                                         $objDrawing->setSrcRect($twoCellAnchor->pic->blipFill->children(Namespaces::DRAWINGML)->srcRect->attributes());
                                                     }
@@ -1597,7 +1603,6 @@ class Xlsx extends BaseReader
                                                     $toOffsetX = Drawing::EMUToPixels($twoCellAnchor->to->colOff);
                                                     $toOffsetY = Drawing::EMUToPixels($twoCellAnchor->to->rowOff);
                                                     $graphic = $twoCellAnchor->graphicFrame->children(Namespaces::DRAWINGML)->graphic;
-                                                    /** @var SimpleXMLElement $chartRef */
                                                     $chartRef = $graphic->graphicData->children(Namespaces::CHART)->chart;
                                                     $thisChart = (string) self::getAttributes($chartRef, $xmlNamespaceBase);
 
@@ -1617,7 +1622,6 @@ class Xlsx extends BaseReader
                                             foreach ($xmlDrawingChildren->absoluteAnchor as $absoluteAnchor) {
                                                 if (($this->includeCharts) && ($absoluteAnchor->graphicFrame)) {
                                                     $graphic = $absoluteAnchor->graphicFrame->children(Namespaces::DRAWINGML)->graphic;
-                                                    /** @var SimpleXMLElement $chartRef */
                                                     $chartRef = $graphic->graphicData->children(Namespaces::CHART)->chart;
                                                     $thisChart = (string) self::getAttributes($chartRef, $xmlNamespaceBase);
                                                     $width = Drawing::EMUToPixels((int) self::getArrayItemString(self::getAttributes($absoluteAnchor->ext), 'cx')[0]);
@@ -1900,7 +1904,6 @@ class Xlsx extends BaseReader
             $value->createText(StringHelper::controlCharacterOOXML2PHP((string) $is->t));
         } elseif ($is !== null) {
             if (is_object($is->r)) {
-                /** @var SimpleXMLElement $run */
                 foreach ($is->r as $run) {
                     if (!isset($run->rPr)) {
                         $value->createText(StringHelper::controlCharacterOOXML2PHP((string) $run->t));